tout bo tout propre

+++ Libft/ft_strlcat.c
@@ -6,13 +6,13 @@
 /*   By: hutricot <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/07 15:32:56 by hutricot          #+#    #+#             */
-/*   Updated: 2018/11/14 11:44:33 by hutricot         ###   ########.fr       */
+/*   Updated: 2018/11/17 12:55:54 by hutricot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcat(char *dest, const char *src, size_t size)
+size_t	ft_strlcat(char *dest, const char *src, size_t size)
 {
 	size_t i;
 	size_t i2;
